Add before_signal and after_signal hooks

+++ circus/watcher.py
@@ -142,7 +142,8 @@ class Watcher(object):
       or the callabled itself and a boolean flag indicating if an
       exception occuring in the hook should not be ignored.
       Possible values for the hook name: *before_start*, *after_start*,
-      *before_spawn*, *before_stop*, *after_stop*.
+      *before_spawn*, *before_stop*, *after_stop*., *before_kill*,
+      *after_kill*, *before_signal* or *after_signal*.
 
     - **options** -- extra options for the worker. All options
       found in the configuration file for instance, are passed
@@ -200,7 +201,8 @@ class Watcher(object):
         self.virtualenv = virtualenv
         self.max_age = int(max_age)
         self.max_age_variance = int(max_age_variance)
-        self.ignore_hook_failure = ['before_stop', 'after_stop']
+        self.ignore_hook_failure = ['before_stop', 'after_stop',
+                                    'before_signal', 'after_signal']
 
         self.respawn = respawn
         self.autostart = autostart
@@ -584,7 +586,14 @@ class Watcher(object):
     def send_signal(self, pid, signum):
         if pid in self.processes:
             process = self.processes[pid]
-            process.send_signal(signum)
+            hook_result = self.call_hook("before_signal",
+                                         pid=pid, signum=signum)
+            if signum != signal.SIGKILL and not hook_result:
+                logger.debug("before_signal hook didn't return True "
+                             "=> signal %i is not sent to %i" % (signum, pid))
+            else:
+                process.send_signal(signum)
+            self.call_hook("after_signal", pid=pid, signum=signum)
         else:
             logger.debug('process %s does not exist' % pid)
 
@@ -724,13 +733,14 @@ class Watcher(object):
             elif wid is None:
                 return slot
 
-    def call_hook(self, hook_name):
+    def call_hook(self, hook_name, **kwargs):
         """Call a hook function"""
-        kwargs = {'watcher': self, 'arbiter': self.arbiter,
-                  'hook_name': hook_name}
+        hook_kwargs = {'watcher': self, 'arbiter': self.arbiter,
+                       'hook_name': hook_name}
+        hook_kwargs.update(kwargs)
         if hook_name in self.hooks:
             try:
-                result = self.hooks[hook_name](**kwargs)
+                result = self.hooks[hook_name](**hook_kwargs)
                 error = None
                 self.notify_event("hook_success",
                                   {"name": hook_name, "time": time.time()})
